% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odwn_structure.R
\name{odwn_structure}
\alias{odwn_structure}
\title{Function defining the structure of the Open Dutch Wordnet}
\usage{
odwn_structure(type)
}
\arguments{
\item{type}{character string as in "/LexicalResource/Lexicon/Synset" or "/LexicalResource/Lexicon/LexicalEntry" or any of these children}
}
\value{
a list with elements attrs and children
}
\description{
Function defining the structure of the Open Dutch Wordnet. There are basically 2 entry points, namely 

Each of these entry points has a field called attrs and children. 

\itemize{
\item "/LexicalResource/Lexicon/Synset": Dutch synthetic sets
\item "/LexicalResource/Lexicon/LexicalEntry": Dutch lexical entries sets
}
}
\examples{
odwn_structure("/LexicalResource/Lexicon/Synset")
odwn_structure("/LexicalResource/Lexicon/LexicalEntry")
odwn_structure("/LexicalResource/Lexicon/LexicalEntry/Sense")
}

